Moves ViewModelBase and RelayCommand to new project, Infrastructure
