+++ source/Infrastructure/Infrastructure.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+	  <TargetFrameworks>net451;netcoreapp3.0</TargetFrameworks>
+  </PropertyGroup>
+
+</Project>
